Take out Command methods for template files processing
Resolve RuboCop offense about class size.
